+++ src/main/kotlin/ris58h/webcalm/javascript/completion/JavaScriptLookupElements.kt
@@ -9,12 +9,22 @@ object JavaScriptLookupElements {
     val WHILE = keywordWithParentheses("while")
     val FOR = keywordWithParentheses("for")
     val SWITCH = keywordWithParentheses("switch")
+    val TRY = keywordWithBlock("try")
+    val CATCH = keywordWithParentheses("catch")
+    val FINALLY = keywordWithBlock("finally")
 
     private fun keywordWithParentheses(keyword: String): LookupElementBuilder {
         return LookupElementBuilder.create("$keyword ()")
             .withInsertHandler(moveCaretRelatively(-1, 0))
             .withPresentableText(keyword).bold()
             .withTailText(" (...)")
+    }
+
+    private fun keywordWithBlock(keyword: String): LookupElementBuilder {
+        return LookupElementBuilder.create("$keyword {\n\n}")
+            .withInsertHandler(moveCaretRelatively(-1, -1))
+            .withPresentableText(keyword).bold()
+            .withTailText(" {...}")
     }
 
     private fun moveCaretRelatively(columnShift: Int, lineShift: Int): InsertHandler<LookupElement> {